PSCI: Add framework to handle composite power states
The state-id field in the power-state parameter of a CPU_SUSPEND call can be
used to describe composite power states specific to a platform. The current PSCI
implementation does not interpret the state-id field. It relies on the target
power level and the state type fields in the power-state parameter to perform
state coordination and power management operations. If composite power states
are used, then the PSCI implementation cannot make global decisions about these
operations since it does not understand the platform specific states.

The extended state-id format support for the power state parameter is
introduced with this patch. It also adds support to involve the platform in
state coordination which facilitates the use of composite power states and
improves the support for entering standby states at multiple power domains.

The PSCI implementation now defines a generic representation of the power-state
parameter. It depends on the platform port to convert the power-state parameter
(possibly encoding a composite power state) passed in a CPU_SUSPEND call to this
representation. It is an array where each index corresponds to a power
level. Each entry contains the local power state the power domain at that power
level could enter.

The platform also defines the macros PLAT_MAX_RET_STATE and PLAT_MAX_OFF_STATE
which lets the PSCI implementation find out which power domains have been
requested to enter a retention or power down state. The PSCI implementation does
not interpret the local power states defined by the platform. The only constraint
is that the PLAT_MAX_RET_STATE < PLAT_MAX_OFF_STATE.

For a power domain tree, an array of local power states is maintained. These are
the states requested for each power domain by all the CPUs which the domain is
an ancestor of. The CPU power level is not considered. During a request to place
multiple power domains in a low power state, the platform is passed an array of
requested power-states for each power domain through the
plat_get_target_pwr_state() API. It coordinates amongst these states to
determine a target local power state for the power domain.

Finally, the platform power management hooks are passed the target local power
states for each power domain using the generic representation described
above. The platform executes operations specific to these target states.

The platform hook for placing a power domain in a standby state
(plat_pm_ops_t.pwr_domain_standby()) should now be used to only place the CPU
power domain in a standby or retention state.

The extended state-id power state format can be enabled by setting the
build flag PSCI_EXTENDED_STATE_ID=1 and it is disabled by default.

NOTE: THE PLATFORM PORTS WILL HAVE TO BE REWORKED TO USE THE NEW FRAMEWORK.

Change-Id: I9d4123d97e179529802c1f589baaa4101759d80c