PSCI: Validate MPIDR in platform_get_core_pos()
This patch changes the behaviour and prototype of the platform_get_core_pos()
API to return an error code (-1) if an invalid MPIDR is passed as the argument.
This enables the caller to safely convert an MPIDR of another CPU to its linear
index without querying the PSCI topology tree e.g. during a call to PSCI CPU_ON.

The mechanism to verify an MPIDR is always platform specific. Hence it is no
longer possible to maintain a weak implementation of this API. It has been made
mandatory for a platform port as a result.

It might not be possible for a platform port to verify an MPIDR before the C
runtime has been setup. This would prevent platform_get_core_pos() from being
callable prior to stack setup. As a result, the generic Trusted Firmware code
does not call this API before the C runtime has been initialised.

The macros/APIs to set CPU data by MPIDR have been removed. Callers now use
either platform_get_core_pos() or platform_my_core_pos() to convert the MPIDR
into a linear index before calling the macro/API to set CPU data by index. Some
of the APIs which manage CPU context have undergone change as well.

This patch is part of a series to remove the assumption in the current PSCI
implementation that MPIDR based affinity levels map directly to levels in a
power domain tree.

NOTE: PLATFORM PORTS SHOULD ENSURE THAT THE IMPLEMENTATIONS OF
PLATFORM_GET_CORE_POS() THEY ARE USING VERIFY THE PASSED MPIDR AND RETURN -1 IN
CASE OF AN ERROR.

Change-Id: I05fc80735c5fababb7ccb942c98411a7aea140ac