PSCI: Add new power domain tree description interface
This patch removes the assumption in the current PSCI implementation that MPIDR
based affinity levels map directly to levels in a power domain tree. The
platform interface for querying the power domain topology has been changed such
that:

1. The generic PSCI code does not generate MPIDRs and use them to query the
   platform about the number of power domains at a particular power level. The
   platform now provides a description of the power domain tree on the SoC
   through a data structure. The existing platform APIs to provide the same
   information have been removed.

2. The linear indices returned by platform_get_core_pos() and
   platform_my_core_pos() are used to retrieve CPU power domain nodes from the
   power domain tree. Power domains above the CPU level are accessed using a
   'parent' field in the tree node descriptors.

The platform describes the power domain tree in an array of 'unsigned
char's. The first entry in the array specifies the number of power domains at
the highest power level implemented in the system. Each susbsequent entry
corresponds to a power domain and contains the number of power domains that are
its direct children.

The PSCI generic code uses this array to populate its internal power domain tree
using the Breadth First Search like algorithm. The tree is split into two
arrays:

1. An array that contains all the CPU power domain nodes

2. An array that contains all the other power domain nodes

A seperate array for CPU nodes allows certain CPU specific optimisations to be
implemented e.g. remove the bakery lock, re-use per-cpu data framework for
storing some information.

Entries in the CPU power domain array are allocated such that the array index of
the domain is equal to the linear index returned by platform_get_core_pos() and
platform_my_core_pos() for the MPIDR corresponding to that domain. This
relationship is key to be able to use an MPIDR to find the corresponding CPU
power domain node, traverse to higher power domain nodes and index into arrays
that contain CPU specific information.

An introductory document has been added to briefly describe the new interface.

NOTE : THIS PATCH BREAKS ALL PLATFORM PORTS. EVERY PLATFORM WILL NEED TO MIGRATE
TO THE NEW MECHANISM FOR DESCRIBING THE POWER DOMAIN TREE OF THE SYSTEM.

Change-Id: I4b444719e8e927ba391cae48a23558308447da13