PSCI: Add API to return index of the calling CPU
This patch introduces a new mandatory platform API: platform_my_core_pos() which
returns the linear index of the calling CPU. This API replaces calls to
platform_get_core_pos() with the MPIDR of the current CPU.

The semantics of platform_get_core_pos() will undergo a change in the next
patch. It will validate the MPIDR passed as an argument. The most frequent use
of this API is to find the linear index of the current CPU. There is no need to
validate the MPIDR in this case. Hence platform_my_core_pos() has been
introduced. Its implementation in this patch simply calls
platform_get_core_pos(). The implementations of these APIs will diverge in the
next patch.

This patch is part of a series to remove the assumption in the current PSCI
implementation that MPIDR based affinity levels map directly to levels in a
power domain tree.

NOTE: THIS PATCH WILL BREAK PLATFORM PORTS. THEY SHOULD PROVIDE A SUITABLE
IMPLEMENTATION OF THIS NEW API.

Change-Id: I89758632b370c2812973a4b2efdd9b81a41f9b69